python/openprinttag-pn5180-shared/pyproject.toml
[project]
name = "openprinttag-pn5180-shared"
dynamic = ["version"]
description = "Shared utilities for OpenPrintTag projects (MQTT, etc.)"
license = "GPL-3.0"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm>=7"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/openprinttag_shared/_version.py"
fallback_version = "0.0.0"
